second_day/flask_backend/requirements.txt
autopep8==1.5.7
backcall==0.2.0
bson==0.5.10
click==8.0.1
decorator==5.1.0
Flask==2.0.1
ipython==7.27.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
numpy==1.21.2
pandas==1.3.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pycodestyle==2.7.0
Pygments==2.10.0
pymongo==3.12.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
toml==0.10.2
traitlets==5.1.0
wcwidth==0.2.5
Werkzeug==2.0.1
